requirements.txt
streamlit
google-cloud
matplotlib
google-cloud-bigquery
json5
db-dtypes
pandas
numpy
plotly